requirements.txt
pyperclip
google-genai
python-dotenv
setproctitle
